requirements.txt
Django==3.0.5
django-cors-headers==3.2.1
django-extensions==3.0.9
django-filter==2.2.0
django-simple-history==2.8.0
django-storages==1.10.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.1.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
Unipath==1.1
django-3-jet==1.0.8
mysqlclient==2.0.3
python-dotenv==0.19.1
boto3==1.18.65